requirements.txt
django
django-nested-inline